dev_requirements.txt
boto3==1.7.30               # AWS client library, same one bundled in lambda environments
Fabric3==1.14.post1         # Task Runner for python
ipdb==0.11                  # Python debugger on steroids
ipython==6.5.0              # Python interpreter on steroids
moto==1.3.3                 # Mocking library for boto
pytest==3.7.1               # Test runner for python
flake8==3.5.0               # static code checking tools
pandas==0.23.4
numpy==1.15.4
